handlers: fix response encoding format
According to the official wire protocol documentation[^1]:

> In simple Query mode, the format of retrieved values is always text,
> except when the given command is a FETCH from a cursor declared with
> the BINARY option.

However, for the binary protocol the situation is more complicated, as
this is determined by the `Bind`[^3] command[^2]:

> Bind also specifies the format to use for any data returned by the
> query; the format can be specified overall, or per-column.

Given the above infromation, this aligns the implementation here with the
`sqllite` and `duckdb` examples from the `pgwire` repository.

To see why the current state is an issue, try running `SELECT 1::int4`
from a Rust crate that uses `rust-postgres`.

[^1]: https://www.postgresql.org/docs/current/protocol-flow.html#PROTOCOL-FLOW-SIMPLE-QUERY
[^2]: https://www.postgresql.org/docs/current/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY
[^3]: https://www.postgresql.org/docs/current/protocol-message-formats.html#PROTOCOL-MESSAGE-FORMATS-BIND